fix typings for selenium standalone
