+++ tests/typings/sync/config.ts
@@ -8,6 +8,19 @@ const conf: WebdriverIO.Config = {
     // can be function only
     afterSuite: () => {},
 
+    services: [
+        ['selenium-standalone', {
+            logs: 'string',
+            installArgs: {
+                version: ''
+            },
+            args: {
+                spawnCb: () => {}
+            },
+            skipSeleniumInstall: true
+        }]
+    ],
+
     transformRequest: (requestOptions) => {
         requestOptions.headers['X-Custom-Auth'] = 'custom_header_value'
         return requestOptions
